% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-make_FACTdata.R
\name{make_FACTdata}
\alias{make_FACTdata}
\title{Make fake FACT-like data to illustrate scoring functions}
\usage{
make_FACTdata(
  namesAC = NULL,
  N = 8,
  AConly = FALSE,
  pmiss1 = 0.1,
  pmiss2 = 0.5,
  misscode = NA
)
}
\arguments{
\item{namesAC}{Character vector of the "Additional Concerns" item names.}

\item{N}{An integer specifying how many fake respondents to generate.
Default is 8.}

\item{AConly}{Logical, if FALSE (the default) the 27 FACT-G items will be
generated in addition to items you supply to \code{namesAC}.  If TRUE, then
only the items you supply to \code{namesAC} are generated.}

\item{pmiss1}{The proportion of items with missing answers in the first half
of respondents.}

\item{pmiss2}{The proportion of items with missing answers in the second half
of respondents.}

\item{misscode}{The value used for missing answers. Default is \code{NA}.
\code{FACTscorer} functions consider values of \code{NA}, \code{8}, and \code{9} to be
missing answers, so these are the only 3 values for this argument useful
for testing \code{FACTscorer} functions.}
}
\value{
A data frame with \code{N} fake respondents (8 by default).  The first half
of respondents will have missing values (coded as \code{9}) for roughly 10\\%
of the items, and the last half will have missing values for roughly 50\\% of
the items.
}
\description{
Generates FACT-like questionnaire answers for fake respondents.
}
\details{
This function is used to generate fake responses to FACT/FACIT
questionnaires.  It is a utility to generate example data for testing the
functions and for demonstrating their usage.  Missing item proportions can be
specified separately for the first and second half of respondents in the
generated data frame (using \code{pmiss1} and \code{pmiss2}, respectively).
If \code{AConly = FALSE}, the default, then FACT-G items will automatically
be generated in addition to any "Additional Concerns" (AC) items given to
\code{namesAC}.  Set \code{AConly = TRUE} to generate only the AC items.  If
\code{namesAC = NULL} (the default), then only FACT-G items will be
generated.

If \code{AConly = TRUE} then at least 2 item names must be given to
\code{namesAC}.
}
\examples{
(exampleDat <- make_FACTdata())
itemNames <- c('B1', 'B2', 'B3', 'B4', 'B5', 'B6', 'B7', 'B8', 'B9', 'P2')
(exampleDat <- make_FACTdata(namesAC = itemNames))
(exampleDat <- make_FACTdata(namesAC = itemNames, AConly = TRUE))
}
