% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx-scoreFACT_AntiA.R
\name{scoreFACT_AntiA}
\alias{scoreFACT_AntiA}
\title{Score the FACT-AntiA (Anti-Angiogenesis)}
\usage{
scoreFACT_AntiA(df, id = NULL, AConly = FALSE, updateItems = FALSE,
  keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame with the questionnaire items, appropriately-named.}

\item{id}{(optional) The quoted name of a variable in \code{df} with a unique
value for each row of \code{df}. If an \code{id} variable is provided
here, it will also be included with the scale scores in the output data
frame. This can facilitate accurate merging of the scale scores back into
the input \code{df}.}

\item{AConly}{(optional) Logical, if omitted or set to \code{FALSE} (the
default) then the function will expect \code{df} to contain the
FACT-General items as well as the more specific "Additional Concerns" (AC)
items. If \code{TRUE}, then the function will only find the AC items in
\code{df}, and will only score the subscale(s) produced by the AC items.}

\item{updateItems}{(optional) Logical, if \code{TRUE} then updated versions
of the items (i.e., re-coded for score calculation) will be returned in the
output data frame with the scale scores. The default, \code{FALSE}, does
not save any updated versions of the items in the resulting data frame.
Most users will want to omit this argument or, equivalently, set it to
\code{FALSE}.}

\item{keepNvalid}{(optional) Logical, if \code{TRUE} then the output data
frame will have additional variables containing the number of valid,
non-missing responses from each respondent to the items on a given scale
(see Details). If \code{FALSE} (the default), these variables will not be
in the returned data frame. Most users will want to omit this argument
or, equivalently, set it to \code{FALSE}.}
}
\value{
A data frame with the following scale scores is returned:

\itemize{
  \item \strong{PWB} - Physical Well-Being subscale
  \item \strong{SWB} - Social/Family Well-Being subscale
  \item \strong{EWB} - Emotional Well-Being subscale
  \item \strong{FWB} - Physical Well-Being subscale
  \item \strong{FACTG} - FACT-G Total Score (PWB+SWB+EWB+FWB)
  \item \strong{AntiAS23} - AntiA subscale, 23-item version
  \item \strong{AntiAS24} - AntiA subscale, 24-item version
  \item \strong{FACT_AntiA_TOTAL} - FACT-AntiA Total Score
  (PWB+SWB+EWB+FWB+AntiAS23)
  \item \strong{FACT_AntiA_TOI} - FACT-AntiA Trial Outcome Index
  (PWB+FWB+AntiAS23)
}
}
\description{
Generates all of the scores of the Functional Assessment of Cancer Therapy -
Anti-Angiogenesis (FACT-AntiA, v4) from item responses.
}
\details{
Given a data frame that includes all of the FACT-AntiA (Version
  4) items as variables, appropriately named, this function generates all of
  the FACT-AntiA scale scores.  It is crucial that the item variables in
  the supplied data frame are named according to FACT conventions.  For
  example, the first physical well-being item should be named GP1, the second
  GP2, and so on. Please refer to the materials provided by
  \url{http://www.facit.org} for the particular questionnaire you are using.
  In particular, refer to the left margin of the official questionnaire
  (i.e., from facit.org) for the appropriate item variable names.

  This questionnaire consists of two components: (1) FACT-G items and (2)
  "Additional Concerns" items.  The FACT-G items (G for General) measure
  general aspects of QoL common to all cancer patients. The "Additional
  Concerns" items measure issues relevant for a specific cancer type,
  treatment, or symptom. These two questionnaire components are typically
  administered together. In some studies, however, ONLY the "Additional
  Concerns" items are administered. The \code{AConly} argument is provided to
  accommodate such cases, and should be set to \code{AConly = TRUE} if ONLY
  the "Additional Concerns" items were administered.

  For more details on the \code{updateItems} and \code{keepNvalid} arguments,
  see the documentation entry for \code{\link{scoreFACTG}} and
  \code{\link{FACTscorer}}.
}
\section{If you get an error message}{

  Keep in mind that this function (and R in general) is case-sensitive.
  Items should be named according to FACT conventions (see \code{Details}
  above).

  All items in \code{df} should be \code{numeric} (i.e., of type
  \code{integer} or \code{double}).

  This function expects missing item responses to be coded as \code{NA},
  \code{8}, or \code{9}, and valid item responses to be coded as \code{0},
  \code{1}, \code{2}, \code{3}, or \code{4}.  Any other value for any of the
  items will result in an error message and no scores.
}

\section{NOTE - Difference Between the 2 Versions of the AntiA Subscale}{


The scoring instructions for the FACT-AntiA on \url{www.FACIT.org} were
updated in February 2016 to include 2 versions of the AntiA Subscale.
\code{AntiAS23} is scored from 23 items and \code{AntiAS24} is scored from 24
items.  The additional item in \code{AntiAS24} is GP2, from the PWB subscale
of the FACT-G.  \code{AntiAS23} is used for the calculation of the FACT-AntiA
Total Score and TOI to avoid counting the GP2 item more than once in those
scores.  An earlier version of the manuscript in the References below
(downloaded from \url{www.FACIT.org}) states: "Because one item ("I have
nausea") from the AntiA subscale is also on the FACT-G, the AntiA subscale
consists of 23 rather than 24 items when combined with the FACT-G."  This
statement appears to be omitted from the official published version of the
manuscript.  This function will provide both versions of the subscale;
however, I assume that most users will want to use \code{AntiAS23} because if
item GP2 was administered it is likely that all of the FACT-G items were
administered.  As of this version of the \code{\link{FACTscorer}} package,
there does not appear to be a version of the questionnaire on
\url{www.FACIT.org} that contains only the 24 items needed for the 24-item
version.  In the validation study in the References, an additional 5
subscales are scored from the 24 items on \code{AntiAS24}.  Currently,
\code{scoreFACT_AntiA} does not produce these subscale scores.
}
\examples{
\dontshow{

## FIRST creating a df with fake item data to score
itemNames <- c('HI7', 'HI12', 'AA1', 'AA2', 'AA3', 'AA4', 'AA5',
'AA6', 'AA7', 'C5', 'AA8', 'BMT13', 'BRM5', 'MS3', 'BRM1', 'Cx6',
'AA9', 'TH5', 'B5', 'AA10', 'GA1', 'B1', 'O2')
exampleDat <- make_FACTdata(namesAC = itemNames)

## NOW scoring the items in exampleDat

## Returns data frame with ONLY scale scores
(scoredDat <- scoreFACT_AntiA(exampleDat))

## Using the id argument (makes merging with original data more fool-proof):
(scoredDat <- scoreFACT_AntiA(exampleDat, id = "ID"))

## Merge back with original data, exampleDat:
mergeDat <- merge(exampleDat, scoredDat, by = "ID")
names(mergeDat)

## If ONLY the "Additional Concerns" items are in df, use AConly = TRUE
(scoredDat <- scoreFACT_AntiA(exampleDat, AConly = TRUE))

## AConly = TRUE with an id variable
(scoredDat <- scoreFACT_AntiA(exampleDat, id = "ID", AConly = TRUE))

## Returns scale scores, plus recoded items (updateItems = TRUE)
## Also illustrates effect of setting keepNvalid = TRUE.
scoredDat <- scoreFACT_AntiA(exampleDat, updateItems = TRUE, keepNvalid = TRUE)
names(scoredDat)
## Descriptives of scored scales
summary(scoredDat[, c('PWB', 'SWB', 'EWB', 'FWB', 'FACTG',
                      'AntiAS23', 'FACT_AntiA_TOTAL', 'FACT_AntiA_TOI')])
}
}
\references{
FACT-AntiA Scoring Guidelines, available at
  \url{http://www.facit.org}

Kaiser, K, Beaumont, JL, Webster, K, Yount, SE, Wagner, LI, Kuzel, TM, &
Cella, D. (2015). Development and validation of the functional assessment of
cancer therapy-antiangiogenesis subscale. \emph{Cancer Medicine}, 4(5),
690-698. \url{http://doi.org/10.1002/cam4.385}
}
\seealso{
This function is very similar to the \code{\link{scoreFACT_B}}
  function.  See the documentation for \code{\link{scoreFACT_B}} for more
  details on the arguments and for examples.  Also see the documentation
  entry for the \code{\link{FACTscorer}} package.  For brevity, examples are
  omitted below, but can be accessed by running
  \code{example(scoreFACT_AntiA)}.
}

