% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca-scoreFACT_V.R
\name{scoreFACT_V}
\alias{scoreFACT_V}
\title{Score the FACT-V}
\usage{
scoreFACT_V(df, id = NULL, AConly = FALSE, updateItems = FALSE,
  keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame with the questionnaire items, appropriately-named.}

\item{id}{(optional) The quoted name of a variable in \code{df} with a unique
value for each row of \code{df}. If an \code{id} variable is provided
here, it will also be included with the scale scores in the output data
frame. This can facilitate accurate merging of the scale scores back into
the input \code{df}.}

\item{AConly}{(optional) Logical, if omitted or set to \code{FALSE} (the
default) then the function will expect \code{df} to contain the
FACT-General items as well as the more specific "Additional Concerns" (AC)
items. If \code{TRUE}, then the function will only find the AC items in
\code{df}, and will only score the subscale(s) produced by the AC items.}

\item{updateItems}{(optional) Logical, if \code{TRUE} then updated versions
of the items (i.e., re-coded for score calculation) will be returned in the
output data frame with the scale scores. The default, \code{FALSE}, does
not save any updated versions of the items in the resulting data frame.
Most users will want to omit this argument or, equivalently, set it to
\code{FALSE}.}

\item{keepNvalid}{(optional) Logical, if \code{TRUE} then the output data
frame will have additional variables containing the number of valid,
non-missing responses from each respondent to the items on a given scale
(see Details). If \code{FALSE} (the default), these variables will not be
in the returned data frame. Most users will want to omit this argument
or, equivalently, set it to \code{FALSE}.}
}
\value{
A data frame with the following scale scores is returned:

\itemize{
  \item \strong{PWB} - Physical Well-Being subscale
  \item \strong{SWB} - Social/Family Well-Being subscale
  \item \strong{EWB} - Emotional Well-Being subscale
  \item \strong{FWB} - Physical Well-Being subscale
  \item \strong{FACTG} - FACT-G Total Score (PWB+SWB+EWB+FWB)
  \item \strong{VCS} - Vulvar Cancer subscale
  \item \strong{FACT_V_TOTAL} - FACT-V Total Score (PWB+SWB+EWB+FWB+VCS)
  \item \strong{FACT_V_TOI} - FACT-V Trial Outcome Index (PWB+FWB+VCS)
}

If \code{AConly = TRUE}, the only scale score returned is \strong{VCS}.

If a variable was given to the \code{id} argument, then that variable will
also be in the returned data frame. Additional, relatively unimportant,
variables will be returned if \code{updateItems = TRUE} or \code{keepNvalid =
TRUE}.
}
\description{
Generates all of the scores of the Functional Assessment of Cancer Therapy - 
Vulvar Cancer (FACT-V, v4) from item responses.
}
\details{
Given a data frame that includes all of the FACT-V (Version
  4) items as variables, appropriately named, this function generates all of
  the FACT-V scale scores.  It is crucial that the item variables in
  the supplied data frame are named according to FACT conventions.  For
  example, the first physical well-being item should be named GP1, the second
  GP2, and so on. Please refer to the materials provided by
  \url{http://www.facit.org} for the particular questionnaire you are using.
  In particular, refer to the left margin of the official questionnaire
  (i.e., from facit.org) for the appropriate item variable names.

  This questionnaire consists of two components: (1) FACT-G items and (2)
  "Additional Concerns" items.  The FACT-G items (G for General) measure
  general aspects of QoL common to all cancer patients. The "Additional
  Concerns" items measure issues relevant for a specific cancer type,
  treatment, or symptom. These two questionnaire components are typically
  administered together. In some studies, however, ONLY the "Additional
  Concerns" items are administered. The \code{AConly} argument is provided to
  accommodate such cases, and should be set to \code{AConly = TRUE} if ONLY
  the "Additional Concerns" items were administered.

  For more details on the \code{updateItems} and \code{keepNvalid} arguments,
  see the documentation entry for \code{\link{scoreFACTG}} and
  \code{\link{FACTscorer}}.
}
\section{If you get an error message}{

  Keep in mind that this function (and R in general) is case-sensitive.
  Items should be named according to FACT conventions (see \code{Details}
  above).

  All items in \code{df} should be \code{numeric} (i.e., of type
  \code{integer} or \code{double}).

  This function expects missing item responses to be coded as \code{NA},
  \code{8}, or \code{9}, and valid item responses to be coded as \code{0},
  \code{1}, \code{2}, \code{3}, or \code{4}.  Any other value for any of the
  items will result in an error message and no scores.
}
\examples{
\dontshow{
## FIRST creating a df with fake item data to score
itemNames <- c('V1', 'V2', 'Cx3', 'V3', 'Cx4', 'V4', 'Cx5', 
'BL4', 'C7', 'Cx6', 'C6', 'BL1', 'V5', 'Cx7', 'V6', 'V7', 'V8', 
'V9', 'HN1')
exampleDat <- make_FACTdata(namesAC = itemNames)

## NOW scoring the items in exampleDat

## Returns data frame with ONLY scale scores
(scoredDat <- scoreFACT_V(exampleDat))

## Using the id argument (makes merging with original data more fool-proof):
(scoredDat <- scoreFACT_V(exampleDat, id = "ID"))

## Merge back with original data, exampleDat:
mergeDat <- merge(exampleDat, scoredDat, by = "ID")
names(mergeDat)

## If ONLY the "Additional Concerns" items are in df, use AConly = TRUE
(scoredDat <- scoreFACT_V(exampleDat, AConly = TRUE))

## AConly = TRUE with an id variable
(scoredDat <- scoreFACT_V(exampleDat, id = "ID", AConly = TRUE))

## Returns scale scores, plus recoded items (updateItems = TRUE)
## Also illustrates effect of setting keepNvalid = TRUE.
scoredDat <- scoreFACT_V(exampleDat, updateItems = TRUE, keepNvalid = TRUE)
names(scoredDat)
## Descriptives of scored scales
summary(scoredDat[, c('PWB', 'SWB', 'EWB', 'FWB', 'FACTG',
                      'VCS', 'FACT_V_TOTAL', 'FACT_V_TOI')])
}
}
\references{
FACT-V Scoring Guidelines, available at
  \url{http://www.facit.org}
  
Janda, M., Obermair, A., Cella, D., Crandon, A.J., & Trimmel, M. (2004).
Vulvar cancer patients' quality of life: A qualitative assessment.
\emph{International Journal of Gynecological Cancer}, 14, 875-881.
}
\seealso{
This function is very similar to the \code{\link{scoreFACT_B}}
  function.  See the documentation for \code{\link{scoreFACT_B}} for more
  details on the arguments and for examples.  Also see the documentation
  entry for the \code{\link{FACTscorer}} package.  For brevity, examples are
  omitted below, but can be accessed by running
  \code{example(scoreFACT_V)}.
}

