% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-tx-scoreFACTGOG_Ntx4.R
\name{scoreFACTGOG_Ntx4}
\alias{scoreFACTGOG_Ntx4}
\title{Score the FACT/GOG-Ntx4}
\usage{
scoreFACTGOG_Ntx4(df, id = NULL, updateItems = FALSE, keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame with the questionnaire items, appropriately-named.}

\item{id}{(optional) The quoted name of a variable in \code{df} with a unique
value for each row of \code{df}. If an \code{id} variable is provided
here, it will also be included with the scale scores in the output data
frame. This can facilitate accurate merging of the scale scores back into
the input \code{df}.}

\item{updateItems}{(optional) Logical, if \code{TRUE} then updated versions
of the items (i.e., re-coded for score calculation) will be returned in the
output data frame with the scale scores. The default, \code{FALSE}, does
not save any updated versions of the items in the resulting data frame.
Most users will want to omit this argument or, equivalently, set it to
\code{FALSE}.}

\item{keepNvalid}{(optional) Logical, if \code{TRUE} then the output data
frame will have additional variables containing the number of valid,
non-missing responses from each respondent to the items on a given scale
(see Details). If \code{FALSE} (the default), these variables will not be
in the returned data frame. Most users will want to omit this argument
or, equivalently, set it to \code{FALSE}.}
}
\value{
A data frame containing the following scale score is returned:

\itemize{
  \item \strong{NtxS4} - FACT/GOG-Ntx-4 subscale
}

If a variable was given to the \code{id} argument, then that variable will
also be in the returned data frame. Additional, relatively unimportant,
variables will be returned if \code{updateItems = TRUE} or \code{keepNvalid
= TRUE}.
}
\description{
Scores the 4-item version of the Functional Assessment of Cancer
Therapy-Neurotoxicity (FACT/GOG-Ntx4) from item responses.
}
\details{
Given a data frame that includes all of the FACT/GOG-Ntx4  items as
  variables, appropriately named, this function generates the FACT/GOG-Ntx4
  scale score.  It is crucial that the item variables in the supplied data
  frame are named according to FACT conventions.  For example, the first item
  should be named GP1, the second GP2, and so on.  Please refer to the
  materials provided by \url{http://www.facit.org} for the particular
  questionnaire you are using. In particular, refer to the left margin of the
  official questionnaire (i.e., from facit.org) for the appropriate item
  variable names.

  For more details on the \code{updateItems} and \code{keepNvalid} arguments,
  see the documentation entry for \code{\link{scoreFACTG}} and
  \code{\link{FACTscorer}}.
}
\section{Note}{

  Keep in mind that this function (and R in general) is
  case-sensitive.

  All items in \code{df} should be \code{numeric} (i.e., of type
  \code{integer} or \code{double}).

  This function expects missing item responses to be coded as \code{NA},
  \code{8}, or \code{9}, and valid item responses to be coded as \code{0},
  \code{1}, \code{2}, \code{3}, or \code{4}.  Any other value for any of the
  items will result in an error message and no scores.
}
\examples{
## FIRST creating a df with fake item data to score
itemNames <- c('Ntx1', 'Ntx2', 'Ntx3', 'Ntx4')
exampleDat <- make_FACTdata(namesAC = itemNames, AConly = TRUE)

## NOW scoring the items in exampleDat

## Returns data frame with ONLY scale score
(scoredDat <- scoreFACTGOG_Ntx4(exampleDat))

## Using the id argument (makes merging with original data more fool-proof):
(scoredDat <- scoreFACTGOG_Ntx4(exampleDat, id = "ID"))

## Merge back with original data, exampleDat:
mergeDat <- merge(exampleDat, scoredDat, by = "ID")
names(mergeDat)

## Returns scale score, plus recoded items (updateItems = TRUE)
## Also illustrates effect of setting keepNvalid = TRUE.
scoredDat <- scoreFACTGOG_Ntx4(exampleDat, updateItems = TRUE, keepNvalid = TRUE)
names(scoredDat)
## Descriptives of scored scales
summary(scoredDat['NtxS4'])
}
\references{
FACT/GOG-Ntx4 Scoring Guidelines, available at
  \url{http://www.facit.org}

Huang HQ, Brady MF, Cella D, Fleming G. Validation and reduction of
FACT/GOG-Ntx subscale for platinum/paclitaxel-induced neurologic symptoms: a
gynecologic oncology group study. \emph{Int J Gynecol Cancer}.
2007;17(2):387-393.
}
\seealso{
For additional details on the function arguments, see
  \code{\link{scoreFACTG}} and \code{\link{FACTscorer}}.  For brevity,
  examples are omitted below.  However, this function is very similar to the
  \code{\link{scoreFACTG7}} function.  See the documentation for
  \code{\link{scoreFACTG7}} for examples.  Alternatively, examples for
  \code{scoreFACTGOG_Ntx4} can be accessed by running
  \code{example(scoreFACTGOG_Ntx4)}.
}

