% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-scoreFANLTC.R
\name{scoreFANLTC}
\alias{scoreFANLTC}
\title{Score the FANLTC}
\usage{
scoreFANLTC(df, id = NULL, updateItems = FALSE, keepNvalid = FALSE)
}
\arguments{
\item{df}{A data frame with the questionnaire items, appropriately-named.}

\item{id}{(optional) The quoted name of a variable in \code{df} with a unique
value for each row of \code{df}. If an \code{id} variable is provided
here, it will also be included with the scale scores in the output data
frame. This can facilitate accurate merging of the scale scores back into
the input \code{df}.}

\item{updateItems}{(optional) Logical, if \code{TRUE} then updated versions
of the items (i.e., re-coded for score calculation) will be returned in the
output data frame with the scale scores. The default, \code{FALSE}, does
not save any updated versions of the items in the resulting data frame.
Most users will want to omit this argument or, equivalently, set it to
\code{FALSE}.}

\item{keepNvalid}{(optional) Logical, if \code{TRUE} then the output data
frame will have additional variables containing the number of valid,
non-missing responses from each respondent to the items on a given scale
(see Details). If \code{FALSE} (the default), these variables will not be
in the returned data frame. Most users will want to omit this argument
or, equivalently, set it to \code{FALSE}.}
}
\value{
A data frame with the following scale scores is returned:

\itemize{
\item \strong{PWB} - Physical Well-Being subscale
\item \strong{SWB} - Social/Family Well-Being subscale
\item \strong{EWB} - Emotional Well-Being subscale
\item \strong{FWB} - Physical Well-Being subscale
\item \strong{FANLTC} - FANLTC Total Score (PWB+SWB+EWB+FWB)
}

If a variable was given to the \code{id} argument, then that variable will
also be in the returned data frame. Additional, relatively unimportant,
variables will be returned if \code{updateItems = TRUE} or \code{keepNvalid =
TRUE}.
}
\description{
Generates all of the scores of the Functional Assessment of
Non-Life Threatening Conditions (FANLTC) from item responses.
}
\details{
Given a data frame that includes all of the FANLTC items as variables,
appropriately named, this function generates all of the FANLTC scale
scores.  It is crucial that the item variables in the supplied data frame
are named according to FACT conventions.  For example, the first physical
well-being item should be named GP1, the second GP2, and so on.  Please
refer to the materials provided by \url{http://www.facit.org} for the
particular questionnaire you are using.  In particular, refer to the left
margin of the official questionnaire (i.e., from facit.org) for the
appropriate item variable names.

Most users will want to omit the \code{updateItems} and \code{keepNvalid}
arguments or, equivalently, set them both to \code{FALSE}. For more details
on these arguments, see the documentation entry for \code{\link{scoreFACTG}}
and \code{\link{FACTscorer}}.
}
\section{Note}{

Keep in mind that this function (and R in general) is case-sensitive.

All variables should be in numeric or integer format.

This scoring function expects missing item responses to be coded as NA,
8, or 9, and valid item responses to be coded as 0, 1, 2, 3, or 4.  Any
other value for any of the items will result in an error message and no
scores.
}

\examples{
\dontshow{
## FIRST creating a df with fake item data to score
exampleDat <- make_FACTdata()

## NOW scoring the items in exampleDat

## Returns data frame with ONLY scale scores
scoredDat <- scoreFANLTC(exampleDat)
scoredDat

## Using the id argument (makes merging with original data more fool-proof):
scoredDat <- scoreFANLTC(exampleDat, id = "ID")
scoredDat
## Merge back with original data, exampleDat:
mergeDat <- merge(exampleDat, scoredDat, by = "ID")
names(mergeDat)

## Returns scale scores, plus recoded items (updateItems = TRUE)
## Also illustrates effect of setting keepNvalid = TRUE.
scoredDat <- scoreFANLTC(exampleDat, updateItems = TRUE, keepNvalid = TRUE)
names(scoredDat)
## Descriptives of scored scales
summary(scoredDat[, c('PWB', 'SWB', 'EWB', 'FWB', 'FANLTC')])
}
}
\references{
FANLTC Scoring Guidelines, available at \url{http://www.facit.org}

Cella, D., & Nowinski, C. J. (2002). Measuring quality of life in chronic
illness: The functional assessment of chronic illness therapy measurement
system. Archives of Physical Medicine and Rehabilitation, 83, Supplement 2,
S10–S17.
}
\seealso{
See \code{\link{FACTscorer}} for more comprehensive information
about using the functions in this package.  This function is very similar
to the \code{\link{scoreFACTG}} function.  See the documentation for
\code{\link{scoreFACTG}} for more details on the arguments and for
examples.  For brevity, examples are omitted below, but can be accessed by
running \code{example(scoreFANLTC)}.
}
